+++ src/nl/vu/cs/bumble/emfcollaborationplugin/handlers/Value.java
@@ -1,0 +1,16 @@
+package nl.vu.cs.bumble.emfcollaborationplugin.handlers;
+
+public class Value {
+	private String $type;
+	
+	public Value() {
+	};
+	
+	public String get$type() {
+		return this.$type;
+	}
+	
+	public void setType(String value) {
+		this.$type = value;
+	}
+}
